refactor : derectory modulization
